refactor: refactor default constants to simplify imports
- Centralize all default constants (HTTP headers, tab indices, UI type names, etc.) into a new constants.py instead of defining them in setting_defaults.py
- Update imports throughout the codebase to pull DEFAULT_HEADERS and UI_TYPENAMES from constants rather than setting_defaults
- Clean up setting_defaults.py by stripping out in-file constant definitions, importing SD_DATA_ROOT from constants, and adding a note about centralized constants management
