+++ scripts/civitai_manager_libs/settings/__init__.py
@@ -43,7 +43,7 @@ config_manager = ConfigManager()
 # Provide default ui_typenames for compatibility with legacy code/tests
 UI_TYPENAMES = getattr(config_manager, 'ui_typenames', None)
 if UI_TYPENAMES is None:
-    from .setting_defaults import DEFAULT_UI_TYPENAMES
+    from .constants import UI_TYPENAMES as DEFAULT_UI_TYPENAMES
 
     UI_TYPENAMES = DEFAULT_UI_TYPENAMES.copy()
     config_manager.settings['ui_typenames'] = UI_TYPENAMES
